+++ travelInsurance/BasePremium.cs
@@ -2,6 +2,6 @@
 {
     public abstract class BasePremium
     {
-        public abstract int GetPremium();
+        public abstract double GetPremium();
     }
 }
